Separate class for Versions API
