+++ whois-api/src/main/java/net/ripe/db/whois/api/rest/WhoisRestService.java
@@ -1,29 +1,16 @@
 package net.ripe.db.whois.api.rest;
 
-import com.google.common.collect.Lists;
 import com.google.common.net.InetAddresses;
 import net.ripe.db.whois.api.QueryBuilder;
-import net.ripe.db.whois.api.rest.domain.WhoisObject;
 import net.ripe.db.whois.api.rest.domain.WhoisResources;
-import net.ripe.db.whois.api.rest.domain.WhoisVersions;
-import net.ripe.db.whois.api.rest.mapper.FormattedServerAttributeMapper;
 import net.ripe.db.whois.api.rest.mapper.WhoisObjectMapper;
-import net.ripe.db.whois.api.rest.mapper.WhoisObjectServerMapper;
-import net.ripe.db.whois.common.Message;
-import net.ripe.db.whois.common.Messages;
 import net.ripe.db.whois.common.dao.RpslObjectDao;
-import net.ripe.db.whois.common.domain.ResponseObject;
 import net.ripe.db.whois.common.rpsl.ObjectType;
 import net.ripe.db.whois.common.rpsl.RpslObject;
 import net.ripe.db.whois.common.source.SourceContext;
 import net.ripe.db.whois.query.QueryFlag;
 import net.ripe.db.whois.query.acl.AccessControlListManager;
-import net.ripe.db.whois.query.domain.DeletedVersionResponseObject;
-import net.ripe.db.whois.query.domain.MessageObject;
 import net.ripe.db.whois.query.domain.QueryException;
-import net.ripe.db.whois.query.domain.VersionResponseObject;
-import net.ripe.db.whois.query.domain.VersionWithRpslResponseObject;
-import net.ripe.db.whois.query.handler.QueryHandler;
 import net.ripe.db.whois.query.query.Query;
 import net.ripe.db.whois.update.domain.Keyword;
 import net.ripe.db.whois.update.domain.Origin;
@@ -51,7 +38,6 @@ import javax.ws.rs.WebApplicationException;
 import javax.ws.rs.core.Context;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
-import java.util.Collections;
 import java.util.List;
 
 import static net.ripe.db.whois.api.rest.RestServiceHelper.getServerAttributeMapper;
@@ -65,10 +51,8 @@ public class WhoisRestService {
     private final RpslObjectDao rpslObjectDao;
     private final RpslObjectStreamer rpslObjectStreamer;
     private final SourceContext sourceContext;
-    private final QueryHandler queryHandler;
     private final AccessControlListManager accessControlListManager;
     private final WhoisObjectMapper whoisObjectMapper;
-    private final WhoisObjectServerMapper whoisObjectServerMapper;
     private final InternalUpdatePerformer updatePerformer;
     private final SsoTranslator ssoTranslator;
     private final LoggerContext loggerContext;
@@ -77,20 +61,16 @@ public class WhoisRestService {
     public WhoisRestService(final RpslObjectDao rpslObjectDao,
                             final RpslObjectStreamer rpslObjectStreamer,
                             final SourceContext sourceContext,
-                            final QueryHandler queryHandler,
                             final AccessControlListManager accessControlListManager,
                             final WhoisObjectMapper whoisObjectMapper,
-                            final WhoisObjectServerMapper whoisObjectServerMapper,
                             final InternalUpdatePerformer updatePerformer,
                             final SsoTranslator ssoTranslator,
                             final LoggerContext loggerContext) {
         this.rpslObjectDao = rpslObjectDao;
         this.rpslObjectStreamer = rpslObjectStreamer;
         this.sourceContext = sourceContext;
-        this.queryHandler = queryHandler;
         this.accessControlListManager = accessControlListManager;
         this.whoisObjectMapper = whoisObjectMapper;
-        this.whoisObjectServerMapper = whoisObjectServerMapper;
         this.updatePerformer = updatePerformer;
         this.ssoTranslator = ssoTranslator;
         this.loggerContext = loggerContext;
@@ -280,88 +260,6 @@ public class WhoisRestService {
         }
     }
 
-    @GET
-    @Produces({MediaType.APPLICATION_XML, MediaType.APPLICATION_JSON})
-    @Path("/{source}/{objectType}/{key:.*}/versions")
-    public Response versions(
-            @Context final HttpServletRequest request,
-            @PathParam("source") final String source,
-            @PathParam("objectType") final String objectType,
-            @PathParam("key") final String key) {
-
-        checkForMainSource(request, source);
-
-        final QueryBuilder queryBuilder = new QueryBuilder()
-                .addCommaList(QueryFlag.SELECT_TYPES, ObjectType.getByName(objectType).getName())
-                .addFlag(QueryFlag.LIST_VERSIONS);
-
-        final Query query = Query.parse(queryBuilder.build(key), Query.Origin.REST, isTrusted(request));
-
-        final VersionsResponseHandler versionsResponseHandler = new VersionsResponseHandler();
-        final int contextId = System.identityHashCode(Thread.currentThread());
-        queryHandler.streamResults(query, InetAddresses.forString(request.getRemoteAddr()), contextId, versionsResponseHandler);
-
-        final List<DeletedVersionResponseObject> deleted = versionsResponseHandler.getDeletedObjects();
-        final List<VersionResponseObject> versions = versionsResponseHandler.getVersionObjects();
-
-        if (versions.isEmpty() && deleted.isEmpty()) {
-            throw new WebApplicationException(Response.status(Response.Status.NOT_FOUND)
-                    .entity(RestServiceHelper.createErrorEntity(request, versionsResponseHandler.getErrors()))
-                    .build());
-        }
-
-        final String type = (versions.size() > 0) ? versions.get(0).getType().getName() : deleted.size() > 0 ? deleted.get(0).getType().getName() : null;
-        final WhoisVersions whoisVersions = new WhoisVersions(type, key, whoisObjectServerMapper.mapVersions(deleted, versions));
-
-        final WhoisResources whoisResources = new WhoisResources();
-        whoisResources.setVersions(whoisVersions);
-        whoisResources.setErrorMessages(RestServiceHelper.createErrorMessages(versionsResponseHandler.getErrors()));
-        whoisResources.includeTermsAndConditions();
-
-        return Response.ok(whoisResources).build();
-    }
-
-    @GET
-    @Produces({MediaType.APPLICATION_XML, MediaType.APPLICATION_JSON})
-    @Path("/{source}/{objectType}/{key:.*}/versions/{version}")
-    public Response version(
-            @Context final HttpServletRequest request,
-            @PathParam("source") final String source,
-            @PathParam("objectType") final String objectType,
-            @PathParam("key") final String key,
-            @PathParam("version") final Integer version) {
-
-        checkForMainSource(request, source);
-
-        final QueryBuilder queryBuilder = new QueryBuilder()
-                .addCommaList(QueryFlag.SELECT_TYPES, ObjectType.getByName(objectType).getName())
-                .addCommaList(QueryFlag.SHOW_VERSION, String.valueOf(version));
-
-        final Query query = Query.parse(queryBuilder.build(key), Query.Origin.REST, isTrusted(request));
-
-        final VersionsResponseHandler versionsResponseHandler = new VersionsResponseHandler();
-        final int contextId = System.identityHashCode(Thread.currentThread());
-        queryHandler.streamResults(query, InetAddresses.forString(request.getRemoteAddr()), contextId, versionsResponseHandler);
-
-        final VersionWithRpslResponseObject versionWithRpslResponseObject = versionsResponseHandler.getVersionWithRpslResponseObject();
-
-        if (versionWithRpslResponseObject == null) {
-            throw new WebApplicationException(Response.status(Response.Status.NOT_FOUND)
-                    .entity(RestServiceHelper.createErrorEntity(request, versionsResponseHandler.getErrors()))
-                    .build());
-        }
-
-        // TODO: [AH] this should use StreamingMarshal to properly handle newlines in errormessages
-        final WhoisResources whoisResources = new WhoisResources();
-        final WhoisObject whoisObject = whoisObjectMapper.map(versionWithRpslResponseObject.getRpslObject(), FormattedServerAttributeMapper.class);
-        whoisObject.setVersion(versionWithRpslResponseObject.getVersion());
-        whoisResources.setWhoisObjects(Collections.singletonList(whoisObject));
-        whoisResources.setErrorMessages(RestServiceHelper.createErrorMessages(versionsResponseHandler.getErrors()));
-        whoisResources.includeTermsAndConditions();
-
-        return Response.ok(whoisResources).build();
-    }
-
     private boolean isTrusted(final HttpServletRequest request) {
         return accessControlListManager.isTrusted(InetAddresses.forString(request.getRemoteAddr()));
     }
@@ -401,45 +299,6 @@ public class WhoisRestService {
 
     private void auditlogRequest(final HttpServletRequest request) {
         loggerContext.log(new HttpRequestMessage(request));
-    }
-
-    private class VersionsResponseHandler extends ApiResponseHandler {
-        final List<VersionResponseObject> versionObjects = Lists.newArrayList();
-        final List<DeletedVersionResponseObject> deletedObjects = Lists.newArrayList();
-        VersionWithRpslResponseObject versionWithRpslResponseObject;
-        private final List<Message> errors = Lists.newArrayList();
-
-        public List<VersionResponseObject> getVersionObjects() {
-            return versionObjects;
-        }
-
-        public List<DeletedVersionResponseObject> getDeletedObjects() {
-            return deletedObjects;
-        }
-
-        public VersionWithRpslResponseObject getVersionWithRpslResponseObject() {
-            return versionWithRpslResponseObject;
-        }
-
-        public List<Message> getErrors() {
-            return errors;
-        }
-
-        @Override
-        public void handle(final ResponseObject responseObject) {
-            if (responseObject instanceof VersionWithRpslResponseObject) {
-                versionWithRpslResponseObject = (VersionWithRpslResponseObject) responseObject;
-            } else if (responseObject instanceof VersionResponseObject) {
-                versionObjects.add((VersionResponseObject) responseObject);
-            } else if (responseObject instanceof DeletedVersionResponseObject) {
-                deletedObjects.add((DeletedVersionResponseObject) responseObject);
-            } else if (responseObject instanceof MessageObject) {
-                Message message = ((MessageObject) responseObject).getMessage();
-                if (message != null && Messages.Type.INFO != message.getType()) {
-                    errors.add(message);
-                }
-            }
-        }
     }
 
     private void checkForMainSource(final HttpServletRequest request, final String source) {
