+++ whois-api/src/main/java/net/ripe/db/whois/api/rest/WhoisServletDeployer.java
@@ -28,6 +28,7 @@ public class WhoisServletDeployer implements ServletDeployer {
 
     private final WhoisRestService whoisRestService;
     private final WhoisSearchService whoisSearchService;
+    private final WhoisVersionService whoisVersionService;
     private final SyncUpdatesService syncUpdatesService;
     private final AsnTransfersRestService asnTransfersRestService;
     private final InetnumTransfersRestService inetnumTransfersRestService;
@@ -44,6 +45,7 @@ public class WhoisServletDeployer implements ServletDeployer {
     @Autowired
     public WhoisServletDeployer(final WhoisRestService whoisRestService,
                                 final WhoisSearchService whoisSearchService,
+                                final WhoisVersionService whoisVersionService,
                                 final SyncUpdatesService syncUpdatesService,
                                 final AsnTransfersRestService asnTransfersRestService,
                                 final InetnumTransfersRestService inetnumTransfersRestService,
@@ -58,6 +60,7 @@ public class WhoisServletDeployer implements ServletDeployer {
                                 final FullTextSearch fullTextSearch) {
         this.whoisRestService = whoisRestService;
         this.whoisSearchService = whoisSearchService;
+        this.whoisVersionService = whoisVersionService;
         this.syncUpdatesService = syncUpdatesService;
         this.asnTransfersRestService = asnTransfersRestService;
         this.inetnumTransfersRestService = inetnumTransfersRestService;
@@ -82,6 +85,7 @@ public class WhoisServletDeployer implements ServletDeployer {
         resourceConfig.register(MultiPartFeature.class);
         resourceConfig.register(whoisRestService);
         resourceConfig.register(whoisSearchService);
+        resourceConfig.register(whoisVersionService);
         resourceConfig.register(syncUpdatesService);
         resourceConfig.register(asnTransfersRestService);
         resourceConfig.register(inetnumTransfersRestService);
