+++ whois-api/src/test/java/net/ripe/db/whois/api/rest/WhoisRestServiceTestIntegration.java
@@ -12,8 +12,6 @@ import net.ripe.db.whois.api.rest.domain.Source;
 import net.ripe.db.whois.api.rest.domain.WhoisObject;
 import net.ripe.db.whois.api.rest.domain.WhoisResources;
 import net.ripe.db.whois.api.rest.domain.WhoisTag;
-import net.ripe.db.whois.api.rest.domain.WhoisVersion;
-import net.ripe.db.whois.api.rest.domain.WhoisVersions;
 import net.ripe.db.whois.api.rest.mapper.DirtyClientAttributeMapper;
 import net.ripe.db.whois.api.rest.mapper.FormattedClientAttributeMapper;
 import net.ripe.db.whois.api.rest.mapper.WhoisObjectMapper;
@@ -100,8 +98,6 @@ import static org.junit.Assert.fail;
 // FIXME: make this into a suite that runs twice: once with XML, once with JSON
 @Category(IntegrationTest.class)
 public class WhoisRestServiceTestIntegration extends AbstractIntegrationTest {
-
-    private static final String VERSION_DATE_PATTERN = "\\d{4}-\\d{2}-\\d{2} \\d{2}:\\d{2}";
 
     private static final RpslObject PAULETH_PALTHEN = RpslObject.parse("" +
             "person:    Pauleth Palthen\n" +
@@ -1425,25 +1421,6 @@ public class WhoisRestServiceTestIntegration extends AbstractIntegrationTest {
         assertThat(response, not(containsString("errormessages")));
     }
 
-    @Test
-    public void lookup_non_streaming_puts_xlink_into_root_element_and_nowhere_else() throws Exception {
-        databaseHelper.addObject(
-                "aut-num:        AS102\n" +
-                "as-name:        End-User-2\n" +
-                "descr:          description\n" +
-                "admin-c:        TP1-TEST\n" +
-                "tech-c:         TP1-TEST\n" +
-                "mnt-by:         OWNER-MNT\n" +
-                "source:         TEST\n");
-
-        final String whoisResources = RestTest.target(getPort(), "whois/test/aut-num/AS102/versions/1")
-                .request(MediaType.APPLICATION_XML)
-                .get(String.class);
-
-        assertThat(whoisResources, containsString("<whois-resources xmlns:xlink=\"http://www.w3.org/1999/xlink\">"));
-        assertThat(whoisResources, containsString("<object type=\"aut-num\" version=\"1\">"));
-        assertThat(whoisResources, containsString("<objects>"));
-    }
 
     @Test
     public void lookup_invalid_suffix_should_not_return_plaintext() {
@@ -3782,262 +3759,6 @@ public class WhoisRestServiceTestIntegration extends AbstractIntegrationTest {
         assertThat(storedObject.getWhoisObjects().get(0).getAttributes(), hasItem(new Attribute("last-modified", "2001-02-14T17:00:00Z")));
         assertThat(storedObject.getWhoisObjects().get(0).getAttributes(), not(hasItem(new Attribute("last-modified", "2001-02-04T17:00:00Z"))));
         assertThat(storedObject.getWhoisObjects().get(0).getAttributes(), hasItem(new Attribute("created", "2001-02-04T17:00:00Z")));
-    }
-
-    // versions
-
-    @Test
-    public void versions_returns_xml() throws IOException {
-        databaseHelper.addObject("" +
-                "aut-num:        AS102\n" +
-                "as-name:        End-User-2\n" +
-                "descr:          description\n" +
-                "admin-c:        TP1-TEST\n" +
-                "tech-c:         TP1-TEST\n" +
-                "mnt-by:         OWNER-MNT\n" +
-                "source:         TEST\n");
-
-        final WhoisResources whoisResources = RestTest.target(getPort(), "whois/test/aut-num/AS102/versions")
-                .request(MediaType.APPLICATION_XML)
-                .get(WhoisResources.class);
-
-        assertThat(whoisResources.getErrorMessages(), is(empty()));
-        final WhoisVersions whoisVersions = whoisResources.getVersions();
-        assertThat(whoisVersions.getType(), is("aut-num"));
-        assertThat(whoisVersions.getKey(), is("AS102"));
-        assertThat(whoisVersions.getVersions(), hasSize(1));
-        final WhoisVersion whoisVersion = whoisVersions.getVersions().get(0);
-        assertThat(whoisVersion, is(new WhoisVersion("ADD/UPD", whoisVersion.getDate(), 1)));
-    }
-
-    @Test
-    public void versions_deleted() throws IOException {
-        final RpslObject autnum = RpslObject.parse("" +
-                "aut-num:        AS102\n" +
-                "as-name:        End-User-2\n" +
-                "descr:          description\n" +
-                "admin-c:        TP1-TEST\n" +
-                "tech-c:         TP1-TEST\n" +
-                "mnt-by:         OWNER-MNT\n" +
-                "source:         TEST\n");
-        databaseHelper.addObject(autnum);
-        databaseHelper.deleteObject(autnum);
-        databaseHelper.addObject(autnum);
-        databaseHelper.updateObject("" +
-                "aut-num:        AS102\n" +
-                "as-name:        End-User-2\n" +
-                "descr:          description\n" +
-                "admin-c:        TP1-TEST\n" +
-                "tech-c:         TP1-TEST\n" +
-                "mnt-by:         OWNER-MNT\n" +
-                "source:         TEST\n");
-
-        final WhoisResources whoisResources = RestTest.target(getPort(), "whois/test/aut-num/AS102/versions")
-                .request(MediaType.APPLICATION_XML)
-                .get(WhoisResources.class);
-        assertThat(whoisResources.getErrorMessages(), is(empty()));
-
-        final List<WhoisVersion> versions = whoisResources.getVersions().getVersions();
-        assertThat(versions, hasSize(3));
-        assertThat(versions.get(0).getDeletedDate(), is(not(nullValue())));
-        assertThat(versions.get(0).getOperation(), is(nullValue()));
-        assertThat(versions.get(0).getDate(), is(nullValue()));
-        assertThat(versions.get(0).getRevision(), is(nullValue()));
-
-        assertThat(versions.get(1).getDeletedDate(), is(nullValue()));
-        assertThat(versions.get(1).getOperation(), is("ADD/UPD"));
-        assertThat(versions.get(1).getRevision(), is(1));
-        assertThat(versions.get(1).getDate(), stringMatchesRegexp(VERSION_DATE_PATTERN));
-
-        assertThat(versions.get(2).getDeletedDate(), is(nullValue()));
-        assertThat(versions.get(2).getOperation(), is("ADD/UPD"));
-        assertThat(versions.get(2).getRevision(), is(2));
-        assertThat(versions.get(2).getDate(), stringMatchesRegexp(VERSION_DATE_PATTERN));
-    }
-
-    @Test
-    public void versions_deleted_versions_json() throws IOException {
-        final RpslObject autnum = RpslObject.parse("" +
-                "aut-num:        AS102\n" +
-                "as-name:        End-User-2\n" +
-                "descr:          description\n" +
-                "admin-c:        TP1-TEST\n" +
-                "tech-c:         TP1-TEST\n" +
-                "mnt-by:         OWNER-MNT\n" +
-                "source:         TEST\n");
-        databaseHelper.addObject(autnum);
-        databaseHelper.deleteObject(autnum);
-        databaseHelper.addObject(autnum);
-        databaseHelper.updateObject("" +
-                "aut-num:        AS102\n" +
-                "as-name:        End-User-2\n" +
-                "descr:          description\n" +
-                "admin-c:        TP1-TEST\n" +
-                "tech-c:         TP1-TEST\n" +
-                "mnt-by:         OWNER-MNT\n" +
-                "source:         TEST\n");
-
-        final WhoisResources whoisResources = RestTest.target(getPort(), "whois/test/aut-num/AS102/versions")
-                .request(MediaType.APPLICATION_JSON)
-                .get(WhoisResources.class);
-        assertThat(whoisResources.getErrorMessages(), is(empty()));
-
-        final List<WhoisVersion> versions = whoisResources.getVersions().getVersions();
-        assertThat(versions, hasSize(3));
-        assertThat(versions.get(0).getDeletedDate(), stringMatchesRegexp(VERSION_DATE_PATTERN));
-        assertThat(versions.get(0).getOperation(), is(nullValue()));
-        assertThat(versions.get(0).getDate(), is(nullValue()));
-        assertThat(versions.get(0).getRevision(), is(nullValue()));
-
-        assertThat(versions.get(1).getDeletedDate(), is(nullValue()));
-        assertThat(versions.get(1).getOperation(), is("ADD/UPD"));
-        assertThat(versions.get(1).getRevision(), is(1));
-        assertThat(versions.get(1).getDate(), stringMatchesRegexp(VERSION_DATE_PATTERN));
-
-        assertThat(versions.get(2).getDeletedDate(), is(nullValue()));
-        assertThat(versions.get(2).getOperation(), is("ADD/UPD"));
-        assertThat(versions.get(2).getRevision(), is(2));
-        assertThat(versions.get(2).getDate(), stringMatchesRegexp(VERSION_DATE_PATTERN));
-    }
-
-    @Test
-    public void versions_last_version_deleted() throws IOException {
-        final RpslObject autnum = RpslObject.parse("" +
-                "aut-num:        AS102\n" +
-                "as-name:        End-User-2\n" +
-                "descr:          description\n" +
-                "admin-c:        TP1-TEST\n" +
-                "tech-c:         TP1-TEST\n" +
-                "mnt-by:         OWNER-MNT\n" +
-                "source:         TEST\n");
-        databaseHelper.addObject(autnum);
-        databaseHelper.deleteObject(autnum);
-
-        final WhoisResources whoisResources = RestTest.target(getPort(), "whois/test/aut-num/AS102/versions")
-                .request(MediaType.APPLICATION_XML)
-                .get(WhoisResources.class);
-        assertThat(whoisResources.getErrorMessages(), is(empty()));
-
-        final List<WhoisVersion> versions = whoisResources.getVersions().getVersions();
-        assertThat(versions, hasSize(1));
-        assertThat(versions.get(0).getDeletedDate(), stringMatchesRegexp(VERSION_DATE_PATTERN));
-        assertThat(versions.get(0).getOperation(), is(nullValue()));
-        assertThat(versions.get(0).getDate(), is(nullValue()));
-        assertThat(versions.get(0).getRevision(), is(nullValue()));
-    }
-
-    @Test(expected = NotFoundException.class)
-    public void versions_no_versions_found() throws IOException {
-        RestTest.target(getPort(), "whois/test/aut-num/AS102/versions")
-                .request(MediaType.APPLICATION_XML)
-                .get(String.class);
-    }
-
-    @Test(expected = NotFoundException.class)
-    public void version_nonexistant_version() throws IOException {
-        databaseHelper.addObject("" +
-                "aut-num:        AS102\n" +
-                "as-name:        End-User-2\n" +
-                "descr:          description\n" +
-                "admin-c:        TP1-TEST\n" +
-                "tech-c:         TP1-TEST\n" +
-                "mnt-by:         OWNER-MNT\n" +
-                "source:         TEST\n");
-
-        RestTest.target(getPort(), "whois/test/aut-num/AS102/versions/2")
-                .request(MediaType.APPLICATION_XML)
-                .get(WhoisResources.class);
-    }
-
-    @Test(expected = NotFoundException.class)
-    public void version_wrong_object_type() throws IOException {
-        databaseHelper.addObject("" +
-                "aut-num:        AS102\n" +
-                "as-name:        End-User-2\n" +
-                "descr:          description\n" +
-                "admin-c:        TP1-TEST\n" +
-                "tech-c:         TP1-TEST\n" +
-                "mnt-by:         OWNER-MNT\n" +
-                "source:         TEST\n");
-
-        RestTest.target(getPort(), "whois/test/inetnum/AS102/versions/1")
-                .request(MediaType.APPLICATION_XML)
-                .get(WhoisResources.class);
-    }
-
-    @Test
-    public void version_returns_xml() throws IOException {
-        final RpslObject autnum = RpslObject.parse("" +
-                "aut-num:        AS102\n" +
-                "as-name:        End-User-2\n" +
-                "descr:          description\n" +
-                "admin-c:        TP1-TEST\n" +
-                "tech-c:         TP1-TEST\n" +
-                "mnt-by:         OWNER-MNT\n" +
-                "source:         TEST\n");
-        databaseHelper.addObject(autnum);
-
-        final WhoisResources whoisResources = RestTest.target(getPort(), "whois/test/aut-num/AS102/versions/1")
-                .request(MediaType.APPLICATION_XML)
-                .get(WhoisResources.class);
-
-        assertThat(whoisResources.getErrorMessages(), is(empty()));
-        assertThat(whoisResources.getWhoisObjects(), hasSize(1));
-        final WhoisObject object = whoisResources.getWhoisObjects().get(0);
-        assertThat(object.getType(), is("aut-num"));
-        assertThat(object.getVersion(), is(1));
-        final List<Attribute> attributes = object.getAttributes();
-        final List<RpslAttribute> originalAttributes = autnum.getAttributes();
-        for (int i = 0; i < originalAttributes.size(); i++) {
-            assertThat(originalAttributes.get(i).getCleanValue().toString(), is(attributes.get(i).getValue()));
-        }
-    }
-
-    @Test
-    public void version_returns_json() throws IOException {
-        final RpslObject autnum = RpslObject.parse("" +
-                "aut-num:        AS102\n" +
-                "as-name:        End-User-2\n" +
-                "descr:          description\n" +
-                "admin-c:        TP1-TEST\n" +
-                "tech-c:         TP1-TEST\n" +
-                "mnt-by:         OWNER-MNT\n" +
-                "source:         TEST\n");
-        databaseHelper.addObject(autnum);
-
-        final WhoisResources whoisResources = RestTest.target(getPort(), "whois/test/aut-num/AS102/versions/1")
-                .request(MediaType.APPLICATION_JSON)
-                .get(WhoisResources.class);
-
-        assertThat(whoisResources.getErrorMessages(), is(empty()));
-        assertThat(whoisResources.getWhoisObjects().size(), is(1));
-        final WhoisObject object = whoisResources.getWhoisObjects().get(0);
-        assertThat(object.getType(), is("aut-num"));
-        assertThat(object.getVersion(), is(1));
-
-        final List<Attribute> attributes = object.getAttributes();
-        final List<RpslAttribute> originalAttributes = autnum.getAttributes();
-        for (int i = 0; i < originalAttributes.size(); i++) {
-            assertThat(originalAttributes.get(i).getCleanValue().toString(), is(attributes.get(i).getValue()));
-        }
-    }
-
-    @Test(expected = NotFoundException.class)
-    public void version_not_showing_deleted_version() throws IOException {
-        final RpslObject autnum = RpslObject.parse("" +
-                "aut-num:        AS102\n" +
-                "as-name:        End-User-2\n" +
-                "descr:          description\n" +
-                "admin-c:        TP1-TEST\n" +
-                "tech-c:         TP1-TEST\n" +
-                "mnt-by:         OWNER-MNT\n" +
-                "source:         TEST\n");
-        databaseHelper.addObject(autnum);
-        databaseHelper.deleteObject(autnum);
-
-        RestTest.target(getPort(), "whois/test/aut-num/AS102/versions/1")
-                .request(MediaType.APPLICATION_XML)
-                .get(WhoisResources.class);
     }
 
     // response format
